requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Django==4.0.3
django-cors-headers==3.11.0
django-daraja==1.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8


dj-database-url==0.5.0
psycopg2-binary==2.9.3
